fix caching, add query devtools
