+++ package.json
@@ -13,6 +13,7 @@
     "@next/font": "^13.1.0",
     "@radix-ui/react-tabs": "1.0.1",
     "@tanstack/react-query": "^4.20.4",
+    "@tanstack/react-query-devtools": "^4.20.4",
     "class-variance-authority": "^0.4.0",
     "clsx": "^1.2.1",
     "framer-motion": "^8.0.2",
