+++ src/pages/vn/[id]/overview.tsx
@@ -1,12 +1,12 @@
 import type { GetServerSideProps } from 'next';
 import { type NextPage } from 'next';
-import { dehydrate, QueryClient } from '@tanstack/react-query';
+import { dehydrate } from '@tanstack/react-query';
 import VnOverview from '../../../features/vn/pages/VnOverviewPage/VnOverviewPage';
 import { overviewQueryOptions } from '../../../features/vn/queries/overview';
+import { queryClient } from '../../../api/queryClient';
 
 /** Get server side props. */
 export const getServerSideProps: GetServerSideProps = async() => {
-  const queryClient = new QueryClient();
   await queryClient.prefetchQuery(overviewQueryOptions);
 
   return {
